- add services page - fix selected menu hoghlighting
